Move `ScalarMaybeUndef` back to rustc
